conduit-backend/requirements.txt
Django==1.10.5
# django-cors-middleware==1.3.1
django-extensions==1.7.1
djangorestframework==3.4.4
PyJWT==1.4.2
six==1.10.0

django-cors-headers==2.4.0

# to read variables from an .env
django-environ==0.4.5

# wsgi application
gunicorn==19.9.0

# for Postgres to function with django
psycopg2-binary==2.8.6
